backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
torchvision
scikit-learn
requests
beautifulsoup4
pandas
tqdm
